Conflicts:
	MVVMAppie/MVVMAppie/MVVMAppie.csproj
	MVVMAppie/MVVMAppie/ProductWindow.xaml
	MVVMAppie/MVVMAppie/ViewModel/MainViewModel.cs
	MVVMAppie/MVVMAppie/ViewModel/SectionVM.cs
	MVVMAppie/MVVMAppie/ViewModel/ViewModelLocator.cs

+++ MVVMAppie/MVVMAppie/Model/Repository/IRepository.cs
@@ -11,7 +11,7 @@ namespace MVVMAppie.Model.Repository
         void Create(T entity);
         void Delete(T entity);
         void Update(T entity);
-        T Get(long Id);
+        T Get(int Id);
         IEnumerable<T> GetAll();
     }
 }
